Moves the --wait test suite to the enable command test suite
